# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.CreateTaskInput:
    properties:
      assignee_id:
        example: 9ec632eb-2345-5678-90bc-def012345678
        type: string
      description:
        example: Token expiration is not handled correctly on the client side
        maxLength: 500
        type: string
      due_date:
        example: "2025-06-30T23:59:59Z"
        type: string
      name:
        example: Fix authentication bug
        maxLength: 100
        type: string
    required:
    - name
    type: object
...
  domain.Task:
    properties:
      assignee_id:
        example: 9ec632eb-2345-5678-90bc-def012345678
        type: string
      column_id:
        example: b3c4d5e6-f7a8-9012-bcde-f12345678901
        type: string
      created_at:
        example: "2025-01-15T09:00:00Z"
        type: string
      description:
        example: Token expiration is not handled correctly on the client side
        type: string
      due_date:
        example: "2025-06-30T23:59:59Z"
        type: string
      id:
        example: c4d5e6f7-a8b9-0123-cdef-123456789012
        type: string
      name:
        example: Fix authentication bug
        type: string
      owner_id:
        example: 7bc4f9d2-1a3e-4c8f-9b2d-5e7a0c4f1d63
        type: string
      updated_at:
        example: "2025-01-15T09:00:00Z"
        type: string
    type: object
...
  domain.UpdateTaskInput:
    properties:
      assignee_id:
        example: 9ec632eb-2345-5678-90bc-def012345678
        type: string
      column_id:
        example: b3c4d5e6-f7a8-9012-bcde-f12345678901
        type: string
      description:
        example: Token expiration is not handled correctly on the client side
        maxLength: 500
        type: string
      due_date:
        example: "2025-06-30T23:59:59Z"
        type: string
      name:
        example: Fix authentication bug
        maxLength: 100
        type: string
    required:
    - column_id
    - name
    type: object
  handler.ErrorResponse:
    properties:
      error:
        example: error description
        type: string
    type: object
...
host: 204.168.173.88:8080
info:
  contact:
    name: Mykyta Kuzminov
    url: https://github.com/mykytakuzminov
  description: |-
    REST API for managing boards, columns and tasks with JWT authentication.
    Supports token-based auth, rate limiting, pagination, sorting and filtering.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Sprintly API
  version: "1.0"
paths:
...
  /columns/{columnID}/tasks:
    get:
      description: Returns all tasks belonging to the specified column. Supports pagination
        and sorting.
      parameters:
      - description: Column ID (UUID)
        in: path
        name: columnID
        required: true
        type: string
      - description: 'Number of results per page (default: 20, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Number of results to skip (default: 0)'
        in: query
        name: offset
        type: integer
      - description: 'Field to sort by: name, due_date, created_at, updated_at (default:
          created_at)'
        in: query
        name: sort
        type: string
      - description: 'Sort direction: ASC or DESC (default: ASC)'
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of tasks
          schema:
            items:
              $ref: '#/definitions/domain.Task'
            type: array
        "400":
          description: Invalid column ID format
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Missing or invalid access token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all tasks in a column
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Creates a new task inside a column. Only the owner of the parent
        board can create tasks.
      parameters:
      - description: Column ID (UUID)
        in: path
        name: columnID
        required: true
        type: string
      - description: Task data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateTaskInput'
      produces:
      - application/json
      responses:
        "201":
          description: Task created successfully
          schema:
            $ref: '#/definitions/domain.Task'
        "400":
          description: Invalid column ID format, request body, or validation error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Missing or invalid access token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Caller is not the board owner
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Column not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a task
      tags:
      - tasks
...
  /tasks/{id}:
    delete:
      description: Deletes a task by ID. Only the task owner (board owner) can perform
        this action.
      parameters:
      - description: Task ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Task deleted successfully
        "400":
          description: Invalid task ID format
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Missing or invalid access token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Caller is not the task owner
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a task
      tags:
      - tasks
    get:
      description: Returns a specific task by its unique identifier.
      parameters:
      - description: Task ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Task data
          schema:
            $ref: '#/definitions/domain.Task'
        "400":
          description: Invalid task ID format
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Missing or invalid access token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get task by ID
      tags:
      - tasks
    patch:
      consumes:
      - application/json
      description: |-
        Updates task fields including column, assignee, name, description and due date.
        The column_id can be used to move the task to another column.
        Only the task owner (board owner) can update it.
      parameters:
      - description: Task ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated task data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateTaskInput'
      produces:
      - application/json
      responses:
        "204":
          description: Task updated successfully
        "400":
          description: Invalid task ID format, request body, or validation error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Missing or invalid access token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Caller is not the task owner
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a task
      tags:
      - tasks
...
  /users/me/tasks:
    get:
      description: Returns all tasks where owner_id matches the authenticated user.
        Supports pagination and sorting.
      parameters:
      - description: 'Number of results per page (default: 20, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Number of results to skip (default: 0)'
        in: query
        name: offset
        type: integer
      - description: 'Field to sort by: name, due_date, created_at, updated_at (default:
          created_at)'
        in: query
        name: sort
        type: string
      - description: 'Sort direction: ASC or DESC (default: ASC)'
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of tasks
          schema:
            items:
              $ref: '#/definitions/domain.Task'
            type: array
        "401":
          description: Missing or invalid access token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all tasks created by current user
      tags:
      - tasks
  /users/me/tasks/assigned:
    get:
      description: Returns all tasks where assignee_id matches the authenticated user.
        Supports pagination and sorting.
      parameters:
      - description: 'Number of results per page (default: 20, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Number of results to skip (default: 0)'
        in: query
        name: offset
        type: integer
      - description: 'Field to sort by: name, due_date, created_at, updated_at (default:
          created_at)'
        in: query
        name: sort
        type: string
      - description: 'Sort direction: ASC or DESC (default: ASC)'
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of assigned tasks
          schema:
            items:
              $ref: '#/definitions/domain.Task'
            type: array
        "401":
          description: Missing or invalid access token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all tasks assigned to current user
      tags:
      - tasks
securityDefinitions:
  BearerAuth:
    description: 'Enter your Bearer token in the format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"